Comments and error
Commented the methods used to communicate with the API with what they do and how they work.
Removed an unused method.
Corrected an error in the recommendation based playlists that would cause the new playlist to contain only songs from the first selected playlist.
Corrected a major error in the concatenate that would cause a loss of tracks.